pyproject.toml
[tool.poetry]
name = "py-script-tool"
version = "0.1.0"
description = ""
authors = ["vuhe <zhuhe202@qq.com>"]
readme = "README.md"

[tool.poetry.scripts]
check-ppcat = "check_ppcat.main:main"
fix-cbz-image = "fix_cbz_image.main:main"
manga-chap-add = "manga_chap_add.main:main"
mkv-ass-subset = "mkv_ass_subset.main:main"

[tool.poetry.dependencies]
python = "^3.12"
rich = "^13.9.4"      # 彩色终端
pillow = "^11.1.0"    # 图片库
otf2ttf = "^0.2"      # 字体文件转换
inquirerpy = "^0.3.4" # 终端交互
selenium = "^4.29.0"
undetected-chromedriver = "^3.5.5"
webdriver-manager = "^4.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
